+++ src/operators/division.java
@@ -1,13 +1,15 @@
+package operators;
+
 import java.util.Scanner;
-public class subtraction {
-    public static int sub(int a,int b){
-        return a-b;
+public class division {
+    public static int div(int a,int b){
+        return a/b;
     }
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int a = sc.nextInt();
         int b = sc.nextInt();
-        int c = sub(a, b);
+        int c = div(a, b);
         System.out.print(c);
     }
 }
