+++ src/operators/subtraction.java
@@ -1,13 +1,15 @@
-import java.util.*;
-public class addition {
-    public static int add(int a,int b){
-        return a+b;
+package operators;
+
+import java.util.Scanner;
+public class subtraction {
+    public static int sub(int a,int b){
+        return a-b;
     }
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int a = sc.nextInt();
         int b = sc.nextInt();
-        int c = add(a, b);
+        int c = sub(a, b);
         System.out.print(c);
     }
 }
